% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estat_census_2020}
\alias{estat_census_2020}
\title{Population of the 2020 census}
\format{
An object of class \code{estat} of length 6.
}
\source{
\url{https://www.e-stat.go.jp/dbview?sid=0003433219}
}
\usage{
estat_census_2020
}
\description{
Population of the 2020 census
}
\keyword{datasets}
