% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estat-collect.R
\name{estat_collect}
\alias{estat_collect}
\title{Get 'e-Stat' data}
\usage{
estat_collect(x, value_name = "value", query = NULL)
}
\arguments{
\item{x}{A \code{estat} object.}

\item{value_name}{A column name of the value.}

\item{query}{A list of additional queries.}
}
\value{
A \code{tbl} of the downloaded data.
}
\description{
Get 'e-Stat' data
}
\examples{
\dontrun{
estat_collect(estat_census_2020)
}

}
