% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estat-download.R
\name{estat_download}
\alias{estat_download}
\title{Download 'e-Stat' data}
\usage{
estat_download(x, value_name = "value", query = NULL)
}
\arguments{
\item{x}{A \code{estat} object.}

\item{value_name}{A column name of the value.}

\item{query}{A list of additional queries.}
}
\value{
A \code{tbl} of the downloaded data.
}
\description{
Download 'e-Stat' data
}
\examples{
\dontrun{
estat_download(estat_census_2020)
}
}
