% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estat-set.R
\name{estat_set}
\alias{estat_set}
\title{Set attributes of 'e-Stat' API}
\usage{
estat_set(i, value)
}
\arguments{
\item{i}{An attribute name.}

\item{value}{An attribute value}
}
\value{
No output..
}
\description{
Set attributes of 'e-Stat' API
}
\examples{
estat_set("estat_limit_downloads", 1e5)
estat_set("estat_limit_items", 1e2)
}
