% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estat.R
\name{estat_set_key}
\alias{estat_set_key}
\title{Set appId of e-Stat API}
\usage{
estat_set_key(appId)
}
\arguments{
\item{appId}{An appId of e-Stat API.}
}
\value{
No output.
}
\description{
Set appId of e-Stat API
}
