% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{estat_table_info}
\alias{estat_table_info}
\title{Get table information for 'e-Stat' data (defunct)}
\usage{
estat_table_info(x)
}
\arguments{
\item{x}{A \code{estat} object.}
}
\value{
This function always errors.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}
}
\details{
\code{estat_table_info()} has been removed. Use \code{\link[=summary]{summary()}} instead.
}
\keyword{internal}
