% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iretp.R
\name{iretp_price}
\alias{iretp_price}
\title{Information on real estate transaction prices API}
\usage{
iretp_price(from, to, pref_code, city_code, lang = c("ja", "en"))
}
\description{
Information on real estate transaction prices API
}
\seealso{
<https://www.land.mlit.go.jp/webland_english/servlet/MainServlet>

<https://www.land.mlit.go.jp/webland/api.html>
}
