% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iretp.R
\name{iretp_trade}
\alias{iretp_trade}
\alias{iretp_city}
\title{Information on real estate transaction prices API}
\usage{
iretp_trade(
  from,
  to,
  pref_code,
  city_code,
  lang = c("ja", "en"),
  .rename = TRUE
)

iretp_city(pref_code, lang = c("ja", "en"), .rename = TRUE)
}
\description{
Information on real estate transaction prices API
}
\seealso{
<https://www.land.mlit.go.jp/webland_english/servlet/MainServlet>

<https://www.land.mlit.go.jp/webland/api.html>
}
