% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/japanstat.R
\docType{package}
\name{japanstat}
\alias{japanstat}
\alias{japanstat-package}
\title{Tools for easy use of 'e-Stat' API}
\description{
Provides tools for using the API of 'e-Stat' (<https://www.e-stat.go.jp/>), a portal site for Japanese government statistics. Includes functions for automatic query generation, data collection and formatting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UchidaMizuki/japanstat}
  \item Report bugs at \url{https://github.com/UchidaMizuki/japanstat/issues}
}

}
\author{
\strong{Maintainer}: Mizuki Uchida \email{uchidamizuki@vivaldi.net}

}
