% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{japanstat_set}
\alias{japanstat_set}
\title{Set attributes}
\usage{
japanstat_set(i, value)
}
\arguments{
\item{i}{An attribute name.}

\item{value}{An attribute value}
}
\value{
No output.
}
\description{
Set attributes
}
\examples{
japanstat_set("estat_limit_collection", 1e5)
japanstat_set("estat_limit_items", 1e2)
}
