% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpstat-package.R
\docType{package}
\name{jpstat}
\alias{jpstat-package}
\alias{jpstat}
\title{Tools for easy use of 'e-Stat' API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides tools to use the 'e-Stat' API (\url{https://www.e-stat.go.jp/}), the portal site for Japanese government statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UchidaMizuki/jpstat}
  \item \url{https://uchidamizuki.github.io/jpstat/}
  \item Report bugs at \url{https://github.com/UchidaMizuki/jpstat/issues}
}

}
\author{
\strong{Maintainer}: Mizuki Uchida \email{uchidamizuki@vivaldi.net}

Authors:
\itemize{
  \item Mizuki Uchida \email{uchidamizuki@vivaldi.net}
}

}
