% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigatr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{activate}
\alias{deactivate}
\alias{rekey}
\alias{rekey_with}
\alias{itemise}
\alias{itemize}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{navigatr}{\code{\link[navigatr:activate]{activate()}}, \code{\link[navigatr:deactivate]{deactivate()}}, \code{\link[navigatr:itemise]{itemise()}}, \code{\link[navigatr:itemize]{itemize()}}, \code{\link[navigatr:rekey]{rekey()}}, \code{\link[navigatr:rekey_with]{rekey_with()}}}
}}

