% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resas.R
\name{resas}
\alias{resas}
\title{Access 'RESAS' data (defunct)}
\usage{
resas(...)
}
\arguments{
\item{...}{Ignored.}
}
\value{
This function always errors.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}
}
\details{
The 'RESAS' API (\url{https://opendata.resas-portal.go.jp/}) was discontinued on
2025-03-24, and accounts as well as API keys were deleted. Therefore
\code{resas()} no longer works and has been made defunct.

As an alternative for prefectural and municipal data, consider the
MLIT Data Platform (DPF) GraphQL API (\url{https://www.mlit-data.jp/}).
}
\keyword{internal}
