% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resas.R
\name{resas_set_apikey}
\alias{resas_set_apikey}
\title{Set 'X_API_KEY' of 'RESAS' API}
\usage{
resas_set_apikey(X_API_KEY)
}
\arguments{
\item{X_API_KEY}{An X_API_KEY of 'RESAS' API.}
}
\description{
Set 'X_API_KEY' of 'RESAS' API
}
