% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estat.R
\name{summary.estat}
\alias{summary.estat}
\alias{summary.tbl_estat}
\title{Summarize the table information of an 'e-Stat' table}
\usage{
\method{summary}{estat}(object, ...)

\method{summary}{tbl_estat}(object, ...)
}
\arguments{
\item{object}{An \code{estat} or \code{tbl_estat} object.}

\item{...}{Ignored.}
}
\value{
A \code{tibble} of table information.
}
\description{
\code{summary()} returns the table-level metadata (title, survey date,
publisher, and so on) of an \code{estat} object created by \code{\link[=estat]{estat()}}.
}
