% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estat.R
\name{tbl_sum.tbl_estat}
\alias{tbl_sum.tbl_estat}
\title{Provide a succinct summary of \code{tbl_estat}}
\usage{
\method{tbl_sum}{tbl_estat}(x)
}
\arguments{
\item{x}{Object to summarise.}
}
\description{
Provide a succinct summary of \code{tbl_estat}
}
