% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estat.R
\name{estat}
\alias{estat}
\title{Get meta-information of 'e-Stat' data}
\usage{
estat(statsDataId, appId = NULL, lang = NULL, query = NULL)
}
\arguments{
\item{statsDataId}{A statistical data ID on 'e-Stat'.}

\item{appId}{An 'appId' of 'e-Stat' API.}

\item{lang}{A language, Japanese (\code{"J"}) or English (\code{"E"}).}

\item{query}{A list of additional queries.}
}
\value{
A \code{estat} object.
}
\description{
The \code{estat} gets the meta-information of a statistical table by using \code{getMetaInfo} of the 'e-Stat' API,
and returns an \code{estat} object that allows editing of meta-information by \code{filter} and \code{select}.
}
\examples{
\dontrun{
estat("https://www.e-stat.go.jp/dbview?sid=0003433219")
}
}
