% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estat-activate.R
\name{estat_activate}
\alias{estat_activate}
\alias{estat_activate_tab}
\alias{estat_activate_time}
\alias{estat_activate_area}
\alias{estat_activate_cat}
\title{Determine which \code{estat} object key to edit.}
\usage{
estat_activate(x, pattern, new_name = NULL)

estat_activate_tab(x, new_name = NULL)

estat_activate_time(x, new_name = NULL)

estat_activate_area(x, new_name = NULL)

estat_activate_cat(x, n, new_name = NULL)
}
\arguments{
\item{x}{A \code{estat} object.}

\item{pattern}{Pattern to look for.}

\item{new_name}{New column name.}

\item{n}{A category number.}
}
\value{
The \code{estat} object which the selected key is active.
}
\description{
Determine which \code{estat} object key to edit.
}
\examples{
estat_activate_tab(estat_census_2020)
estat_activate_cat(estat_census_2020, 1)
estat_activate_area(estat_census_2020)
estat_activate_time(estat_census_2020)
}
