% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estat.R
\name{estat_table_info}
\alias{estat_table_info}
\title{Get table information for 'e-Stat' data}
\usage{
estat_table_info(x)
}
\arguments{
\item{x}{A \code{estat} object.}
}
\value{
A \code{tbl} of the table information.
}
\description{
Get table information for 'e-Stat' data
}
