% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iretp.R
\name{iretp_trade}
\alias{iretp_trade}
\alias{iretp_city}
\title{Information on real estate transaction prices API}
\usage{
iretp_trade(
  from,
  to,
  pref_code,
  city_code,
  lang = c("ja", "en"),
  .rename = TRUE
)

iretp_city(pref_code, lang = c("ja", "en"), .rename = TRUE)
}
\arguments{
\item{from}{A start time.}

\item{to}{An end time.}

\item{pref_code}{A prefecture code.}

\item{city_code}{A city code.}

\item{lang}{Language.}

\item{.rename}{Change the names?}
}
\description{
Information on real estate transaction prices API
}
\seealso{
\url{https://www.land.mlit.go.jp/webland_english/servlet/MainServlet}

\url{https://www.land.mlit.go.jp/webland/api.html}
}
