% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpstat-package.R
\docType{package}
\name{jpstat}
\alias{jpstat}
\alias{jpstat-package}
\title{Tools for easy use of 'e-Stat' API}
\description{
Provides tools to use APIs such as e-Stat (\url{https://www.e-stat.go.jp/}), the portal site for Japanese government statistics, and RESAS (Regional Economy and Society Analyzing System, \url{https://resas.go.jp}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UchidaMizuki/jpstat}
  \item \url{https://uchidamizuki.github.io/jpstat/}
  \item Report bugs at \url{https://github.com/UchidaMizuki/jpstat/issues}
}

}
\author{
\strong{Maintainer}: Mizuki Uchida \email{uchidamizuki@vivaldi.net}

}
