% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigatr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{activate}
\alias{deactivate}
\alias{rekey}
\alias{rekey_with}
\alias{itemise}
\alias{itemize}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{navigatr}{\code{\link[navigatr]{activate}}, \code{\link[navigatr:activate]{deactivate}}, \code{\link[navigatr]{itemise}}, \code{\link[navigatr:itemise]{itemize}}, \code{\link[navigatr]{rekey}}, \code{\link[navigatr:rekey]{rekey_with}}}
}}

