% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R, R/resas.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{filter}
\alias{select}
\alias{slice}
\alias{tbl_format_header}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{slice}}}

  \item{pillar}{\code{\link[pillar]{tbl_format_header}}}
}}

