% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R, R/navigatr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{select}
\alias{collect}
\alias{activate}
\alias{deactivate}
\alias{rekey}
\alias{rekey_with}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:compute]{collect}}, \code{\link[dplyr]{select}}}

  \item{navigatr}{\code{\link[navigatr]{activate}}, \code{\link[navigatr:activate]{deactivate}}, \code{\link[navigatr]{rekey}}, \code{\link[navigatr:rekey]{rekey_with}}}
}}

