% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resas.R
\name{resas}
\alias{resas}
\title{Access 'RESAS' data}
\usage{
resas(X_API_KEY, path)
}
\arguments{
\item{X_API_KEY}{An 'X-API-KEY' of 'RESAS' API.}

\item{path}{A 'RESAS' API path.}
}
\value{
A \code{resas} object.
}
\description{
Access 'RESAS' data
}
\seealso{
\url{https://opendata.resas-portal.go.jp/}
}
