% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resas.R
\name{resas}
\alias{resas}
\title{Access 'RESAS' data}
\usage{
resas(
  X_API_KEY = deprecated(),
  path,
  query = list(),
  to_snakecase = TRUE,
  names_sep = "/",
  rectangle = TRUE
)
}
\arguments{
\item{X_API_KEY}{(Deprecated) an 'X-API-KEY' of 'RESAS' API.}

\item{path}{A 'RESAS' API path.}

\item{query}{Additional queries.}

\item{to_snakecase}{Whether the parameters and responses should be named as
snake cases or not?}

\item{names_sep}{A character that separates the names of the responses.}

\item{rectangle}{Whether to rectangle the data or not?}
}
\value{
A \code{resas} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
\url{https://opendata.resas-portal.go.jp/}
}
